requirements.txt
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git
openai
python-dotenv